requirements.txt
livekit>=1.0.9
# livekit-agents[openai]>=1.1.3,<2.0.0
# livekit-agents[deepgram]>=1.1.3,<2.0.0
# livekit-agents[silero]>=1.1.3,<2.0.0
# livekit-agents[turn-detector]>=1.1.3,<2.0.0
livekit-agents[openai,silero,deepgram,turn-detector]>=1.1.1,<2.0.0
types-protobuf<5

python-dotenv~=1.0
av>=14.0.0
cachetools>=6.0.0
pyyaml>=6.0.1
grpcio>=1.73.0
protobuf>=6.0.0
jinja2>=3.1.0
